requirements.txt
apache-airflow==2.8.1
huntflow-api-client==2.0.1
python-dotenv==1.0.1
numpy==1.26.3
psycopg2-binary==2.9.9
selenium==4.18.1
beautifulsoup4==4.12.3
pandas==2.2.1
pytest==8.2.1
pylint==3.2.3
black==24.2.0